+++ src/version3/models/evaluateMetaData.ts
@@ -1,0 +1,8 @@
+import { JiraExpressionsComplexityBean } from './jiraExpressionsComplexityBean';
+import { JExpEvaluateIssuesMetaBean } from './jExpEvaluateIssuesMetaBean';
+
+/** Contains information about the expression evaluation. */
+export interface EvaluateMetaData {
+  complexity?: JiraExpressionsComplexityBean;
+  issues?: JExpEvaluateIssuesMetaBean;
+}