+++ src/version3/parameters/evaluateJSISJiraExpression.ts
@@ -1,0 +1,12 @@
+import { JiraExpressionEvaluateRequestBean } from '../models';
+
+export interface EvaluateJSISJiraExpression extends JiraExpressionEvaluateRequestBean {
+  /**
+   * Use [expand](#expansion) to include additional information in the response. This parameter accepts
+   * `meta.complexity` that returns information about the expression complexity. For example, the number of expensive
+   * operations used by the expression and how close the expression is to reaching the [complexity
+   * limit](https://developer.atlassian.com/cloud/jira/platform/jira-expressions/#restrictions). Useful when designing
+   * and debugging your expressions.
+   */
+  expand?: string;
+}