+++ src/version3/parameters/index.ts
@@ -1,3 +1,7 @@
+export * from './getNotificationSchemeForProject';
+export * from './updateWorkflowSchemeMappings';
+export * from './updateSchemes';
+export * from './readWorkflowSchemes';
 export * from './evaluateJSISJiraExpression';
 export * from './replaceCustomFieldOption';
 export * from './getCustomFieldsConfigurations';