+++ src/version3/parameters/updateSchemes.ts
@@ -1,0 +1,3 @@
+import { WorkflowSchemeUpdateRequest } from '../models';
+
+export interface UpdateSchemes extends WorkflowSchemeUpdateRequest {}