+++ src/version3/parameters/updateWorkflowSchemeMappings.ts
@@ -1,0 +1,3 @@
+import { WorkflowSchemeUpdateRequiredMappingsRequest } from '../models';
+
+export interface UpdateWorkflowSchemeMappings extends WorkflowSchemeUpdateRequiredMappingsRequest {}